🐛 산책 수정시 버그 수정
- 선택 해제 불가능하던 버그 수정
- 상세페이지 삭제 안되던 버그 수정

+++ FrontEnd/lib/components/diary/walk_detail_form.dart
@@ -58,8 +58,10 @@ class WalkDetailsPage extends StatelessWidget {
                       onPressed: () {
                         // debugPrint("삭제");
                         // 스크린샷 삭제
-                        File(imagePath).deleteSync();
 
+                        if (File(imagePath).existsSync()) {
+                          File(imagePath).deleteSync();
+                        }
                         // 삭제 요청 전송
                         apiWalk.deleteWalk(walkId);
                         // 뒤로가기
